Consolidate Dukascopy durable functions
